pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bcarver"
version = "1.3.0"
requires-python = ">=3.8"
description = "file carving utility based on a set of file carving patterns"
readme = {file = "README.md", content-type = "text/markdown"}
dependencies = [
    "PyYAML==6.0.3",
    "tqdm==4.67.3"
]

[project.scripts]
bcarver = "bcarver:main"
